% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeMp3.R
\name{merge_mp3}
\alias{merge_mp3}
\title{Merge Multiple Audio Files into a Single MP3}
\usage{
merge_mp3()
}
\value{
No return value, called for side effects.
}
\description{
This function merges multiple audio files (MP3, WAV, AAC, FLAC, OGG, M4A) into a single MP3 file using FFmpeg.
}
\examples{
\dontrun{
if (interactive()) {
  merge_mp3()
}
}

}
