% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_app.R
\name{run_converter_app}
\alias{run_converter_app}
\title{Launch the Video to MP3 Converter App}
\usage{
run_converter_app()
}
\value{
No return value, called for side effects.
}
\description{
This function converts video files to MP3. It provides a simple interface for selecting an input folder
and processing the files, with the converted MP3 files saved in the same input folder.
}
\examples{
\dontrun{
if (interactive()) {
  run_converter_app()
}
}

}
