% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trim_Audio.R
\name{trim_audio}
\alias{trim_audio}
\title{Trim an Audio File}
\usage{
trim_audio()
}
\value{
No return value, called for side effects.
}
\description{
This function trims an audio file based on user-provided start and end times using FFmpeg.
}
\examples{
\dontrun{
if (interactive()) {
  trim_audio()
}
}

}
