% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LJmp3converter-package.R
\docType{package}
\name{LJmp3converter-package}
\alias{LJmp3converter}
\alias{LJmp3converter-package}
\title{LJmp3converter: Convert Video Files to MP3 Format using 'FFmpeg'}
\description{
Converts video files to 'MP3' using 'FFmpeg', which is dynamically downloaded to avoid bundling any third-party binaries. Users must ensure compliance with the license terms of 'FFmpeg' when using the package. See \url{https://github.com/BtbN/FFmpeg-Builds/releases/download/latest/ffmpeg-master-latest-win64-gpl.zip} for details.
}
\author{
\strong{Maintainer}: Lijin Arakkandathil Thekkathil \email{lijin5673@gmail.com}

}
\keyword{internal}
